Move the methods to app.rb
