+++ data/authors.json
@@ -1,0 +1,7 @@
+[
+  {
+    "id": 419,
+    "first_name": "kha",
+    "last_name": null
+  }
+]
